+++ integreat_cms/xliff/utils.py
@@ -34,7 +34,7 @@ download_storage = FileSystemStorage(
 logger = logging.getLogger(__name__)
 
 
-def pages_to_xliff_file(request, pages, target_language, only_public=False):
+def pages_to_xliff_file(request, pages, target_languages, only_public=False):
     """
     Export a list of page IDs to a ZIP archive containing XLIFF files for a specified target language (or just a single
     XLIFF file if only one page is converted)
@@ -45,8 +45,8 @@ def pages_to_xliff_file(request, pages, target_language, only_public=False):
     :param pages: list of pages which should be translated
     :type pages: list [ ~integreat_cms.cms.models.pages.page.Page ]
 
-    :param target_language: The target language (should not be the region's default language)
-    :type target_language: :class:`~integreat_cms.cms.models.languages.language.Language`
+    :param target_languages: list of target languages (should exclude the region's default language)
+    :type target_languages: [ :class:`~integreat_cms.cms.models.languages.language.Language` ]
 
     :param only_public: Whether only public versions should be exported
     :type only_public: bool
@@ -55,22 +55,25 @@ def pages_to_xliff_file(request, pages, target_language, only_public=False):
     :rtype: str
     """
     logger.debug(
-        "XLIFF export started by %r for pages %r and language %r",
+        "XLIFF export started by %r for pages %r and languages %r",
         request.user,
         pages,
-        target_language,
+        target_languages,
     )
     # Generate unique directory for this export
     dir_name = uuid.uuid4()
     # Create XLIFF files
     xliff_paths = {}
-    for page in pages:
-        try:
-            xliff_paths[page] = page_to_xliff(
-                page, target_language, dir_name, only_public=only_public
-            )
-        except RuntimeError as e:
-            messages.error(request, e)
+    for target_language in target_languages:
+        for page in pages:
+            try:
+                xliff_paths[f"{target_language}/{page}"] = page_to_xliff(
+                    page, target_language, dir_name, only_public=only_public
+                )
+            except RuntimeError as e:
+                messages.error(request, e)
+            except RuntimeWarning as e:
+                messages.warning(request, e)
     # Check how many XLIFF files were created
     if len(xliff_paths) == 0:
         return None
@@ -90,7 +93,11 @@ def pages_to_xliff_file(request, pages, target_language, only_public=False):
     # Generate file path for ZIP archive
     timestamp = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M")
     region = pages[0].region
-    zip_name = f"{region.slug}_{timestamp}_{region.get_source_language(target_language.slug).slug}_{target_language.slug}.zip"
+    if len(target_languages) == 1:
+        target_language = target_languages[0]
+        zip_name = f"{region.slug}_{timestamp}_{region.get_source_language(target_language.slug).slug}_{target_language.slug}.zip"
+    else:
+        zip_name = f"{region.slug}_{timestamp}_multiple_languages.zip"
     actual_filename = download_storage.save(f"{dir_name}/{zip_name}", ContentFile(""))
     # Create ZIP archive
     create_zip_archive(
@@ -122,7 +129,9 @@ def page_to_xliff(page, target_language, dir_name, only_public=False):
     :param only_public: Whether only public versions should be exported
     :type only_public: bool
 
-    :raises RuntimeError: If the selected page translation does not have a source translation
+    :raises RuntimeWarning: If the selected page translation does not have a source translation
+
+    :raises RuntimeError: When an unexpected error occurs during serialization
 
     :return: The path of the generated XLIFF file
     :rtype: str
@@ -145,16 +154,20 @@ def page_to_xliff(page, target_language, dir_name, only_public=False):
     if not source_translation:
         source_language = page.region.get_source_language(target_language.slug)
         logger.warning(
-            "Page translation %r does not have a source translation in %r and therefore cannot be exported to XLIFF.",
+            "Page translation %r does not have a %ssource translation in %r and therefore cannot be exported to XLIFF.",
             target_page_translation,
+            "published " if only_public else "",
             source_language,
         )
-        raise RuntimeError(
+        raise RuntimeWarning(
             _(
-                "Page {} does not have a source translation in {} and "
+                "Page {} does not have a {}source translation in {} and "
                 "therefore cannot be exported as XLIFF for translation to {}."
             ).format(
-                target_page_translation.readable_title, source_language, target_language
+                target_page_translation.readable_title,
+                _("published") + " " if only_public else "",
+                source_language,
+                target_language,
             )
         )